feat(fgdpp,toolpp): move FGD parser/writer into toolpp library
