--- src/fgdpp/_fgdpp.cmake
@@ -1,4 +1,0 @@
-add_pretty_parser(fgdpp
-        SOURCES
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/fgdpp/fgdpp.h"
-        "${CMAKE_CURRENT_LIST_DIR}/fgdpp.cpp")
