+++ test/res/toolpp/cmdseq/binary.wc
