+++ test/res/toolpp/cmdseq/keyvalues.wc
