+++ test/res/toolpp/fgd/game/base.fgd
