+++ test/res/toolpp/fgd/game/halflife2.fgd
