+++ test/res/toolpp/fgd/game/portal.fgd
