+++ test/res/toolpp/fgd/game/portal2.fgd
